Refactor CLI with lazy subcommands and deferring imports (#1920)
* try delaying import + lazy subcommands


* lazy group initial draft

* fix lint

* add tests and fix lint

* add default command and tests

* add default command and tests

* address PR comments

---------
