+++ package/kedro_viz/launchers/cli/__init__.py
@@ -1,0 +1,1 @@
+"""`kedro_viz.launchers.cli` launches the viz server as a CLI app."""
