+++ package/tests/test_launchers/test_utils.py
@@ -1,17 +1,15 @@
 from pathlib import Path
 from unittest import mock
-from unittest.mock import Mock, call, patch
+from unittest.mock import Mock
 
 import pytest
 import requests
 
-from kedro_viz.constants import VIZ_DEPLOY_TIME_LIMIT
 from kedro_viz.launchers.utils import (
     _check_viz_up,
     _find_kedro_project,
     _is_project,
     _start_browser,
-    viz_deploy_progress_timer,
 )
 
 
@@ -54,24 +52,6 @@ def test_check_viz_up(host, port, status_code, expected_result, mocker):
 
     result = _check_viz_up(host, port)
     assert result == expected_result
-
-
-def test_viz_deploy_progress_timer(capsys):
-    mock_process_completed = Mock()
-    mock_process_completed.value = 0
-
-    with patch("kedro_viz.launchers.utils.sleep") as mock_sleep:
-        viz_deploy_progress_timer(mock_process_completed, VIZ_DEPLOY_TIME_LIMIT)
-
-    assert mock_sleep.call_count == VIZ_DEPLOY_TIME_LIMIT + 1
-
-    expected_sleep_calls = [call(1)] * (VIZ_DEPLOY_TIME_LIMIT + 1)
-    mock_sleep.assert_has_calls(expected_sleep_calls)
-    captured = capsys.readouterr()
-
-    for second in range(1, VIZ_DEPLOY_TIME_LIMIT + 1):
-        expected_output = f"...Creating your build/deploy Kedro-Viz ({second}s)"
-        assert expected_output in captured.out
 
 
 class TestIsProject:
